Add app_template.rs to sov-modules-stf-template
